requirements.txt
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
nltk==3.3
pytz==2018.7
six==1.11.0
SpeechRecognition==3.8.1
